Move widget classes to components/Widget
